Language update (#202)
I'm not sure this is perfect but thought I'd make an attempt. Hopefully my intention is clear in trying to give the end user some context as far as how this work seeing that it doesn't magically build your sass with gulp without you actually adding the command and creating a gulpfile, etc.