Update filesystem.settings.php for ACSF (#411)
I'm not 100% sure if this is the best way to address this. Previously, my project (as far as I could tell) was not setting "file_public_path", so I'm assuming it was set maybe through ACSF configuration? The latest release of BLT on L#8 sets the "file_public_path" to the default for all sites, even SF. To fix this, I had to explicitly set the file path after this file loads. Wondering if this change should be incorporated here upstream? Again, I'm not sure if this is the appropriate place for this, but this was my best guess.