Prompt for deploy properties
Adds property prompts so users don't have to remember the syntax for passing options to `blt deploy`. This still allows for properties to be passed in via the `-D` option (so CI can still deploy without input) but for local deploys a user doesn't need to add additional options.

The if check needs to remain as the `propertyprompt` unless a default value is added as the target allows you to skip entering a value.