Fix error message when security-updates test fails
The error message says that `disable-targets.tests.security-updates` must be set to false if I want to disable security checks. This doesn't work. After searching around a bit, I found this comment: https://github.com/acquia/blt/issues/1440#issuecomment-297451659. It seems that it should be set to true (and it makes sense, because the key name is `disable-targets`).