Issue #2796: Fixes failing deploys due to config status check. (#2802)
Fixes #2796 

This Drush command needs to use whatever Drush context is provided, especially the URI in the case of site factory. Really, all of the Drush calls in this file should probably be refactored to be more consistent (like we do with the Drush task for setup and similar commands), but this solves the immediate problem.

I'm not sure if it's the best approach, but I've confirmed it works on ACSF and doesn't break anything locally.